requirements.txt
beautifulsoup4==4.6.0
blinker==1.4
certifi==2017.11.5
chardet==3.0.4
click==6.7
Flask==0.12.2
Flask-SQLAlchemy==2.3.2
gunicorn==19.7.1
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
PyMySQL==0.7.11
pytest==3.2.3
PyYAML==3.12
raven==6.3.0
requests==2.18.4
SQLAlchemy==1.1.15
urllib3==1.22
Werkzeug==0.12.2
